# Fix running performance benchmarks.